requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
Flask-SocketIO>=5.3
eventlet>=0.35
python-dotenv>=1.0.0
gunicorn==21.2.0
